Extract light header and brightness slider components from light card to new components
